add stylings
